feat: support streaming ssr
