+++ packages/server/core/src/base/libs/warmup.ts
@@ -1,10 +1,9 @@
 export function warmup(bundles: Array<string | undefined>) {
   bundles.forEach(bundle => {
-    try {
-      bundle && import(bundle);
-    } catch (_) {
-      // ignore errors, if we can't import the bundle
-      // such as: the bundle is not exists.
-    }
+    bundle &&
+      import(bundle).catch(_ => {
+        // ignore errors, if we can't import the bundle
+        // such as: the bundle is not exists.
+      });
   });
 }
